kvflowcontrol: move tokens per stream to base pkg
Move `kvadmission.flow_controller.(elastic|regular)_tokens_per_stream`
into the parent flow control package, `kvflowcontrol`. This is necessary
to share the setting between the v1 and v2 implementation.

Epic: none
Release note: None
