+++ pkg/kv/kvserver/kvflowcontrol/kvflowcontroller/BUILD.bazel
@@ -12,7 +12,6 @@ go_library(
         "//pkg/base",
         "//pkg/kv/kvserver/kvflowcontrol",
         "//pkg/kv/kvserver/kvflowcontrol/kvflowinspectpb",
-        "//pkg/settings",
         "//pkg/settings/cluster",
         "//pkg/util/admission/admissionpb",
         "//pkg/util/buildutil",
